tests/submission/MOSQUITO/device-orientation/event-provides-event-handler-idl-attribute-ondevicemotion.html

<!DOCTYPE html>
<html>

  <head>
	<title>User agents must also provide an event handler IDL attribute HTML5 named ondevicemotion on the window object:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  </head>

  <body>

  
  		<h1>User agents must also provide an event handler IDL attribute HTML5 named ondevicemotion on the window object:</h1>
    <div id="log"></div>
	
	<script type="text/javascript">
		t1 = async_test("User agents provide an event handler IDL attribute HTML5 named ondevicemotion");
		t2 = async_test("The type of this event handler must be 'DeviceMotionEvent'");
		ename = null;
		etype = null;
		window.ondevicemotion = function(e){
			ename = e.type;
			etype = ""+e.constructor
		};
		
		setTimeout("testing()", 200);
		
		function testing(){
			t1.step(function() {assert_equals(ename, "devicemotion")});
			t2.step(function() {assert_true(etype.lastIndexOf("DeviceMotionEvent") != -1)});
			t1.done();
			t2.done();
		}
	</script>
	
  </body>
  
</html>
